+++ HttpShare.Windows/DataContexts/MainWindowDataContext.cs
@@ -5,7 +5,7 @@ using System.ComponentModel;
 
 using HttpShare.Windows.Controls;
 
-namespace HttpShare.Windows.Models;
+namespace HttpShare.Windows.DataContexts;
 
 /// <summary>
 /// The data context class for <see cref="MainWindow"/>. 
@@ -41,4 +41,17 @@ public sealed class MainWindowDataContext : INotifyPropertyChanged
 	/// </summary>
 	public string ServerToggleButtonText => IsServerRunning ? "Stop Server" : "Start Server";
 
+
+	private bool enableServerToggleButton = true;
+
+	public bool EnableServerToggleButton
+	{
+		get => enableServerToggleButton;
+
+		set
+		{
+			enableServerToggleButton = value;
+			PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(EnableServerToggleButton)));
+		}
+	}
 }