+++ HttpShare/Files/IInboxFile.cs
@@ -1,0 +1,6 @@
+namespace HttpShare.Files;
+
+public interface IInboxFile : IServerFile
+{
+	public string SenderName { get; }
+}