+++ HttpShare/Files/IServerFile.cs
@@ -1,0 +1,10 @@
+namespace HttpShare.Files;
+
+public interface IServerFile
+{
+	public string Name { get; }
+
+	public byte[] Data { get; }
+
+	public long Length { get; }
+}