+++ HttpShare/Sessions/DualSession.cs
@@ -1,18 +1,19 @@
 // Copyright 2024 Hilde801 (https://github.com/hilde801)
 // This file is a part of http-share
 
+using HttpShare.Files;
+
 namespace HttpShare.Sessions;
 
 /// <summary>
 /// Describes a server session that supports both sending and receiving files. 
 /// </summary>
-public sealed class DualSession(ICollection<OutboxFile> outboxFiles) : ServerSession,
-	ISendSession, IReceiveSession
+public sealed class DualSession(IEnumerable<IOutboxFile> outboxFiles) : ServerSession, ISendSession, IReceiveSession
 {
 	/// <summary>
 	///  A collection of files to the sent to client devices.
 	/// </summary>
-	public ICollection<OutboxFile> OutboxFiles => outboxFiles;
+	public IEnumerable<IOutboxFile> OutboxFiles => outboxFiles;
 
 	/// <summary>
 	/// The implementation of <see cref="IReceiveSession.OnReceivedFiles"/>.
@@ -24,7 +25,7 @@ public sealed class DualSession(ICollection<OutboxFile> outboxFiles) : ServerSes
 	/// Invokes <see cref="OnReceivedFiles"/>.
 	/// </summary>
 	/// <param name="files">Files received from client devices.</param>
-	public void InvokeReceivedFilesEvent(ICollection<InboxFile> files)
+	public void InvokeReceivedFilesEvent(ICollection<IInboxFile> files)
 	{
 		OnReceivedFiles?.Invoke(files);
 	}