+++ HttpShare/Sessions/ISendSession.cs
@@ -1,6 +1,8 @@
 // Copyright 2024 Hilde801 (https://github.com/hilde801)
 // This file is a part of http-share
 
+using HttpShare.Files;
+
 namespace HttpShare.Sessions;
 
 /// <summary>
@@ -11,5 +13,5 @@ public interface ISendSession
 	/// <summary>
 	/// A collection of <see cref="OutboxFile"/>s to be sent to connected clients.
 	/// </summary>
-	public ICollection<OutboxFile> OutboxFiles { get; }
+	public IEnumerable<IOutboxFile> OutboxFiles { get; }
 }